Requests using PSR7

+++ src/Controller/DeleteVideoController.php
@@ -5,6 +5,9 @@ namespace Alura\Mvc\Controller;
 use Alura\Mvc\Entity\Video;
 use Alura\Mvc\Repository\VideoRepository;
 use Alura\Mvc\Helper\FlashMessageTrait;
+use Nyholm\Psr7\Response;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
 
 class DeleteVideoController implements Controller {
 
@@ -14,16 +17,21 @@ class DeleteVideoController implements Controller {
 
     }
 
-    public function handle() {
-        $id = $_GET['id'] ?? null;
+    public function handle(ServerRequestInterface $request): ResponseInterface {
+        $queryParams = $request->getQueryParams();
+        $id = filter_var($queryParams['id'], FILTER_SANITIZE_STRING);
+        if(!$id) {
+            $this->addErrorMessage('ID inválido.');
+            return new Response(302, ['Location' => '/']);
+        }
 
         $video = $this->repository->find($id);
 
         if (!$video || $this->repository->remove($id) === false) {
             $this->addErrorMessage('Não foi possível excluir o vídeo.');
-            header('Location: /');
+            return new Response(302, ['Location' => '/']);
         } else {
-            header('Location: /');
+            return new Response(302, ['Location' => '/']);
         }
     }
